Fixed deque
the q&a mentioned that the deque could have issues where you would get a
null pointer exception if you empty the list and start filling it again,
so i fixed that. Now we have comleted code for Deque and
RandomizedQueue. All thats left to do now is to create the client
program and write java doc